refactor: split query bindings and authorization
