+++ crates/jazz/src/node/query_eval.rs
@@ -86,31 +86,6 @@ pub(crate) fn exact_known_state_declaration_for_test(
 pub(crate) const JAZZ_APP_ROWS_SINK: &str = "app_rows";
 const PENDING_BINDING_SOURCE_SHAPE: &str = "__jazz_pending_binding_source";
 
-/// The caller-owned meaning of an absent claim while binding a prepared plan.
-/// Ordinary prepared queries require all declared bindings. Authorization
-/// support, on the other hand, must represent an absent policy claim as an
-/// empty proof for the commit's permission subject.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum PreparedClaimBindingMode {
-    Strict,
-    FailClosedAuthorizationSupport,
-}
-
-/// Exact, action-specific policy support compiled for a hypothetical operation.
-///
-/// The support key deliberately excludes the row/candidate operation key: two
-/// operations reuse hydration only when their compiled support is identical.
-#[cfg_attr(not(test), allow(dead_code))]
-pub(crate) struct AuthorizationSupportScope {
-    pub(crate) key: AuthorizationSupportScopeKey,
-    pub(crate) operation: AuthorizationOperationKey,
-    /// The sole read/serving semantics under which support can authorize an
-    /// operation.  A scope must never be satisfied by a branch, snapshot, or
-    /// local-tier view that merely happens to have the same query identity.
-    pub(crate) options: RegisterShapeOptions,
-    pub(crate) subscriptions: Vec<(ValidatedQuery, Binding)>,
-}
-
 /// Aggregate terminal membership is structurally identified by the aggregate
 /// query plan and its synthetic group-key member. Its table label is not an
 /// identity and must not participate in public delivery decisions.
@@ -291,98 +266,6 @@ where
     /// `RegisterShape` / `Subscribe` a subscriber sent over a connection.
     pub(crate) fn registered_shape(&self, shape_id: ShapeId) -> Option<ValidatedQuery> {
         self.query.registered_shapes.get(&shape_id).cloned()
-    }
-
-    fn program_binding_for_shape(
-        &self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        source_shape: Option<String>,
-        extra_user_params: BTreeMap<String, ColumnType>,
-        claim_params: BTreeMap<String, ProgramClaimParam>,
-    ) -> ProgramBinding {
-        let mut param_types = shape.params().clone();
-        param_types.extend(extra_user_params.clone());
-        ProgramBinding {
-            id: binding.binding_id(),
-            source_shape,
-            extra_user_params,
-            param_types,
-            claim_params,
-            values: binding.values().clone(),
-        }
-    }
-
-    fn program_binding_for_shape_and_policy(
-        &self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        source_shape: Option<String>,
-        extra_user_params: BTreeMap<String, ColumnType>,
-        claim_params: BTreeMap<String, ProgramClaimParam>,
-        policy: &PolicyContext,
-    ) -> Result<ProgramBinding, Error> {
-        self.program_binding_for_shape_and_policy_with_prepared_claim_mode(
-            shape,
-            binding,
-            source_shape,
-            extra_user_params,
-            claim_params,
-            policy,
-            PreparedClaimBindingMode::Strict,
-        )
-    }
-
-    fn program_binding_for_shape_and_policy_with_prepared_claim_mode(
-        &self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        source_shape: Option<String>,
-        extra_user_params: BTreeMap<String, ColumnType>,
-        claim_params: BTreeMap<String, ProgramClaimParam>,
-        policy: &PolicyContext,
-        prepared_claim_binding_mode: PreparedClaimBindingMode,
-    ) -> Result<ProgramBinding, Error> {
-        let mut program_binding = self.program_binding_for_shape(
-            shape,
-            binding,
-            source_shape,
-            extra_user_params,
-            claim_params.clone(),
-        );
-        if !claim_params.is_empty() {
-            let mut values = binding.values().clone();
-            for (name, claim) in &claim_params {
-                let Some(value) = prepared_claim_value(&claim.path, policy)? else {
-                    if prepared_claim_binding_mode
-                        == PreparedClaimBindingMode::FailClosedAuthorizationSupport
-                    {
-                        return Err(Error::AuthorizationSupportMissingClaim(
-                            claim.path.0.join("."),
-                        ));
-                    }
-                    // An absent claim cannot establish a policy proof. Leave
-                    // it as a capability gap so the policy source resolver
-                    // lowers the proof to an empty authorization graph.
-                    if matches!(policy, PolicyContext::AuthorizationSubplan { .. }) {
-                        return Err(Error::QueryCapability(format!(
-                            "policy authorization requires unbound claim {}",
-                            claim.path.0.join(".")
-                        )));
-                    }
-                    return Err(Error::InvalidStoredValue(
-                        "claim prepared param is not bound",
-                    ));
-                };
-                values.insert(
-                    name.clone(),
-                    coerce_prepared_binding_value(value, &claim.ty),
-                );
-            }
-            program_binding.id = binding_id_for_values(&values);
-            program_binding.values = values;
-        }
-        Ok(program_binding)
     }
 
     fn compile_current_query_program(
@@ -799,161 +682,6 @@ where
         Ok(rows)
     }
 
-    fn policy_authorization_row_id_graph(
-        &mut self,
-        request: QueryProgramRequest,
-    ) -> Result<PolicyAuthorizationGraph, Error> {
-        self.query_engine_read_metrics.policy_authorization_graphs += 1;
-        let cache_key = policy_authorization_graph_cache_key(&request);
-        if let Some(graph) = self.query.policy_authorization_graph_cache.get(&cache_key) {
-            return Ok(graph.clone());
-        }
-        let proof_table = match &request.policy {
-            PolicyContext::AuthorizationSubplan {
-                protected_source, ..
-            } => Some(protected_source.table.clone()),
-            PolicyContext::System | PolicyContext::Identity { .. } => None,
-        };
-        if let Some(table) = &proof_table {
-            let depth = self.query.policy_proof_stack.len();
-            if self
-                .query
-                .policy_proof_stack
-                .iter()
-                .any(|active| active == table)
-            {
-                return Err(Error::PolicyProofCycle {
-                    table: table.clone(),
-                    depth,
-                });
-            }
-            self.query.policy_proof_stack.push(table.clone());
-        }
-
-        let result = (|| {
-            let program = self.compile_query_program_request(request)?;
-            let graph = lowered_terminal_graph(&program, "policy.authorized_rows")?;
-            let route_fields = program
-                .lowered
-                .terminals
-                .iter()
-                .find_map(|terminal| {
-                    (terminal.sink == "policy.authorized_rows").then(|| match &terminal.output {
-                        OutputTerminalSchema::Fact(fact) => {
-                            output_routing_fields_for_query_eval(fact)
-                        }
-                        OutputTerminalSchema::AppRows(_) => BTreeSet::new(),
-                    })
-                })
-                .unwrap_or_default();
-            let graph = PolicyAuthorizationGraph {
-                graph,
-                route_fields,
-            };
-            self.query
-                .policy_authorization_graph_cache
-                .insert(cache_key, graph.clone());
-            Ok(graph)
-        })();
-
-        if proof_table.is_some() {
-            self.query
-                .policy_proof_stack
-                .pop()
-                .expect("policy proof stack entry is balanced");
-        }
-        result
-    }
-
-    pub(super) fn branch_read_policy_authorized_branch_ids(
-        &mut self,
-        branch_id: BranchId,
-        identity: AuthorId,
-    ) -> Result<BTreeSet<RowUuid>, Error> {
-        let Some(policy) = self.catalogue.schema.branch_read_policy.clone() else {
-            return Ok(BTreeSet::from([RowUuid(branch_id.0)]));
-        };
-        let mut query = policy;
-        query.filters.push(crate::query::eq(
-            crate::query::col("id"),
-            crate::query::lit(Value::Uuid(branch_id.0)),
-        ));
-        let policy_shape = query.validate(&self.catalogue.schema)?;
-        let policy_binding = policy_shape.bind(BTreeMap::new())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            &self.catalogue.schema,
-            ParamBindingMode::InlineAllReachableSeeds,
-        )?;
-        if !policy_shape.params().is_empty() {
-            return Err(Error::QueryCapability(
-                "branch read policy filters with runtime parameters must lower through query-engine binding sources"
-                    .to_owned(),
-            ));
-        }
-        let binding = policy_shape.bind(BTreeMap::new())?;
-        let input_shape = self.normalized_row_set_shape(&policy_shape, &binding)?;
-        let input = RowSetProgramInput {
-            binding: self.program_binding_for_shape(
-                &policy_shape,
-                &binding,
-                None,
-                BTreeMap::new(),
-                binding_claim_params_for_shape(&input_shape, policy_shape.params()),
-            ),
-            shape: input_shape,
-        };
-        let request = QueryProgramRequest {
-            authorization_mode: QueryAuthorizationMode::TrustedServing,
-            reads: current_query_read_set(
-                &input.shape,
-                policy_shape.schema_version(),
-                policy_shape.schema_version(),
-                DurabilityTier::Local,
-                None,
-            ),
-            policy: match self.query_program_policy_context(identity) {
-                PolicyContext::Identity {
-                    mode,
-                    permission_subject,
-                    claims,
-                    attribution,
-                } => PolicyContext::AuthorizationSubplan {
-                    protected_source: root_source_id(policy_shape.query().table.as_str()),
-                    role: PolicyDecisionRole::Read,
-                    mode,
-                    permission_subject,
-                    claims,
-                    attribution,
-                },
-                other => other,
-            },
-            input,
-            output: current_query_output_request(
-                CurrentQueryProgramOutput::AuthorizedRows,
-                policy_shape.query(),
-            ),
-        };
-        let graph = self.policy_authorization_row_id_graph(request)?.graph;
-        let deltas = self.database.query_graph(graph).map_err(Error::Groove)?;
-        let row_idx =
-            deltas
-                .descriptor
-                .field_index("row_uuid")
-                .ok_or(Error::InvalidStoredValue(
-                    "branch read authorization terminal is missing row_uuid",
-                ))?;
-        let mut rows = BTreeSet::new();
-        for (record, weight) in deltas.iter() {
-            if weight <= 0 {
-                continue;
-            }
-            rows.insert(RowUuid(record.get_uuid(row_idx)?));
-        }
-        Ok(rows)
-    }
-
     fn current_query_program_request(
         &self,
         shape: &ValidatedQuery,
@@ -1078,287 +806,6 @@ where
             input,
             output: current_query_output_request(output, shape.query()),
         })
-    }
-
-    fn query_program_policy_context(&self, identity: AuthorId) -> PolicyContext {
-        if identity == AuthorId::SYSTEM {
-            PolicyContext::System
-        } else {
-            let mut claims = default_policy_claim_values(identity);
-            if let Some(session_claims) = self.session_claims.get(&identity) {
-                claims.extend(session_claims.clone());
-            }
-            claims.insert("sub".to_owned(), Value::Uuid(identity.0));
-            PolicyContext::Identity {
-                mode: PolicyEnforcementMode::Enforcing,
-                permission_subject: identity,
-                claims,
-                attribution: None,
-            }
-        }
-    }
-
-    pub(super) fn write_policy_query_allows_current_row(
-        &mut self,
-        policy: &crate::query::Query,
-        row_uuid: RowUuid,
-        identity: AuthorId,
-    ) -> Result<bool, Error> {
-        let mut query = policy.clone();
-        query.filters.push(crate::query::eq(
-            crate::query::col("id"),
-            crate::query::lit(Value::Uuid(row_uuid.0)),
-        ));
-        let policy_shape = query.validate(&self.catalogue.schema)?;
-        let policy_binding = policy_shape.bind(BTreeMap::new())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            &self.catalogue.schema,
-            ParamBindingMode::InlineAllReachableSeeds,
-        )?;
-        let binding = policy_shape.bind(BTreeMap::new())?;
-        let program = self.compile_current_query_program_with_selected_access_paths(
-            &policy_shape,
-            &binding,
-            DurabilityTier::Local,
-            identity,
-            CurrentQueryProgramOutput::AppRows,
-        )?;
-        self.write_policy_query_program_allows(&program, &policy_shape, &binding)
-    }
-
-    /// Authorize an inline old/candidate row through the query program.
-    ///
-    /// Insert candidates reinterpret plain `inherits(parent)` as parent
-    /// update-using authorization. Existing/update-check rows retain ordinary
-    /// read inheritance unless the policy names an explicit write operation.
-    pub(super) fn write_policy_query_allows_candidate(
-        &mut self,
-        table: &TableSchema,
-        policy: &crate::query::Query,
-        row_uuid: RowUuid,
-        cells: &BTreeMap<String, Value>,
-        identity: AuthorId,
-        insert_candidate: bool,
-        branch_id: Option<BranchId>,
-    ) -> Result<bool, Error> {
-        let policy_schema_version = if self
-            .catalogue
-            .schema
-            .tables
-            .iter()
-            .any(|known| known == table)
-        {
-            self.catalogue.current_schema_version_id
-        } else {
-            self.catalogue
-                .catalogue_schemas
-                .iter()
-                .find_map(|(schema_version, payload)| {
-                    payload
-                        .schema
-                        .tables
-                        .iter()
-                        .any(|known| known == table)
-                        .then_some(*schema_version)
-                })
-                .unwrap_or(self.catalogue.current_schema_version_id)
-        };
-        self.write_policy_query_allows_candidate_for_schema(
-            policy_schema_version,
-            table,
-            policy,
-            row_uuid,
-            cells,
-            identity,
-            insert_candidate,
-            branch_id,
-        )
-    }
-
-    pub(super) fn write_policy_query_allows_candidate_for_schema(
-        &mut self,
-        policy_schema_version: SchemaVersionId,
-        table: &TableSchema,
-        policy: &crate::query::Query,
-        row_uuid: RowUuid,
-        cells: &BTreeMap<String, Value>,
-        identity: AuthorId,
-        insert_candidate: bool,
-        branch_id: Option<BranchId>,
-    ) -> Result<bool, Error> {
-        let mut policy = policy.clone();
-        if insert_candidate {
-            for inherits in &mut policy.inherits {
-                if inherits.operation == crate::query::InheritsOperation::Select {
-                    inherits.operation = crate::query::InheritsOperation::Update;
-                }
-            }
-            for branch in &mut policy.policy_branches {
-                for inherits in &mut branch.inherits {
-                    if inherits.operation == crate::query::InheritsOperation::Select {
-                        inherits.operation = crate::query::InheritsOperation::Update;
-                    }
-                }
-            }
-        }
-        let policy_schema = if policy_schema_version == self.catalogue.current_schema_version_id {
-            &self.catalogue.schema
-        } else {
-            &self
-                .catalogue
-                .catalogue_schemas
-                .get(&policy_schema_version)
-                .ok_or(Error::InvalidStoredValue("policy schema payload missing"))?
-                .schema
-        };
-        let policy_shape = policy
-            .clone()
-            .validate_with_schema_version(policy_schema, policy_schema_version)?;
-        let policy_binding = policy_shape.bind(BTreeMap::new())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            policy_schema,
-            ParamBindingMode::InlineAllReachableSeeds,
-        )?;
-        let binding = policy_shape.bind(BTreeMap::new())?;
-        let input_shape = self.normalized_row_set_shape(&policy_shape, &binding)?;
-        let root_source = root_source_id(policy_shape.query().table.as_str());
-        let input = RowSetProgramInput {
-            binding: self.program_binding_for_shape(
-                &policy_shape,
-                &binding,
-                query_binding_source_shape_for_parts_if_needed(
-                    policy_shape.params(),
-                    &binding_claim_params_for_shape(&input_shape, policy_shape.params()),
-                ),
-                BTreeMap::new(),
-                binding_claim_params_for_shape(&input_shape, policy_shape.params()),
-            ),
-            shape: input_shape,
-        };
-        let policy = match self.query_program_policy_context(identity) {
-            PolicyContext::Identity {
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            } => PolicyContext::AuthorizationSubplan {
-                protected_source: root_source_id(policy_shape.query().table.as_str()),
-                role: PolicyDecisionRole::Write,
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            },
-            other => other,
-        };
-        let request = QueryProgramRequest {
-            authorization_mode: QueryAuthorizationMode::TrustedServing,
-            reads: match branch_id {
-                Some(branch_id) => branch_query_read_set(
-                    &input.shape,
-                    policy_shape.schema_version(),
-                    DurabilityTier::Local,
-                    branch_id,
-                ),
-                None => current_query_read_set(
-                    &input.shape,
-                    policy_shape.schema_version(),
-                    policy_shape.schema_version(),
-                    DurabilityTier::Local,
-                    None,
-                ),
-            },
-            policy,
-            input,
-            output: current_query_output_request(
-                CurrentQueryProgramOutput::PolicyPredicate,
-                policy_shape.query(),
-            ),
-        };
-        let candidate = current_row_from_cells(table, row_uuid, cells)?;
-        let inline_sources = BTreeMap::from([(root_source, vec![candidate])]);
-        let access_paths = if branch_id.is_some() {
-            BTreeMap::new()
-        } else {
-            self.current_query_primary_key_access_paths(&policy_shape, &binding)?
-        };
-        let program = self.compile_query_program_request_with_inline_sources_and_access_paths(
-            request,
-            inline_sources,
-            access_paths,
-        )?;
-        self.write_policy_query_program_allows(&program, &policy_shape, &binding)
-    }
-
-    pub(super) fn branch_write_policy_query_allows_candidate(
-        &mut self,
-        branch_id: BranchId,
-        table: &TableSchema,
-        policy: &crate::query::Query,
-        row_uuid: RowUuid,
-        cells: &BTreeMap<String, Value>,
-        identity: AuthorId,
-        insert_candidate: bool,
-    ) -> Result<bool, Error> {
-        self.write_policy_query_allows_candidate(
-            table,
-            policy,
-            row_uuid,
-            cells,
-            identity,
-            insert_candidate,
-            Some(branch_id),
-        )
-    }
-
-    fn write_policy_query_program_allows(
-        &mut self,
-        program: &QueryProgram,
-        policy_shape: &ValidatedQuery,
-        binding: &Binding,
-    ) -> Result<bool, Error> {
-        let deltas =
-            match self.prepared_query_plan_from_program(&program, &policy_shape, &binding)? {
-                PreparedQueryPlan::Graph(graph) => {
-                    self.database.query_graph(graph).map_err(Error::Groove)?
-                }
-                PreparedQueryPlan::Prepared { shape, params } => {
-                    let values = binding_values_for_plan(
-                        &binding,
-                        &params,
-                        &program.request.policy,
-                        PreparedClaimBindingMode::Strict,
-                    )?;
-                    take_required_sink_deltas(
-                        self.bind_shape_snapshot(shape, &values)?,
-                        JAZZ_APP_ROWS_SINK,
-                    )?
-                }
-                PreparedQueryPlan::PeerMaintainedMarker => {
-                    return Err(Error::InvalidStoredValue(
-                        "peer maintained marker cannot execute write policy plan",
-                    ));
-                }
-            };
-        Ok(deltas.iter().any(|(_, weight)| weight > 0))
-    }
-
-    /// Evaluate a validated query shape against this node's local knowledge.
-    ///
-    /// Phase B step 2 returns output-relation rows only. Provenance-closure
-    /// shipping and settled result set reads are introduced by the wire step.
-    pub fn query_rows(
-        &mut self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        tier: DurabilityTier,
-    ) -> Result<Vec<CurrentRow>, Error> {
-        self.require_catalogue_ready()?;
-        self.query_rows_with_prepared_plan(shape, binding, tier, None)
     }
 
     pub(crate) fn query_rows_with_prepared_plan(
@@ -3174,771 +2621,6 @@ where
         self.database.unsubscribe(subscription_id);
         snapshot
     }
-
-    fn policy_filtered_current_source_graph_via_query_engine(
-        &mut self,
-        policy_request: Result<QueryProgramRequest, Error>,
-        base: GraphBuilder,
-        output_fields: &[String],
-    ) -> Result<PolicyAuthorizationGraph, Error> {
-        self.query_engine_read_metrics
-            .policy_authorized_source_joins += 1;
-        let policy_request = match policy_request {
-            Ok(policy_request) => policy_request,
-            Err(Error::QueryCapability(err)) if err.contains("PolicyProofCycle") => {
-                return Err(Error::QueryCapability(err));
-            }
-            Err(Error::QueryCapability(_)) => {
-                return Ok(empty_policy_filtered_current_source_graph(
-                    base,
-                    output_fields,
-                ));
-            }
-            Err(err) => return Err(err),
-        };
-        // The protected storage source has no binding fields of its own, but
-        // the authorization proof is routed by the enclosing prepared
-        // binding. Carry that descriptor alongside the source before joining
-        // the proof so a later storage delta has every route field the proof
-        // advertises.
-        let binding_routes = policy_request
-            .input
-            .binding
-            .source_shape
-            .as_ref()
-            .map(|shape| {
-                let descriptor = RecordDescriptor::new(
-                    policy_request
-                        .input
-                        .binding
-                        .param_types
-                        .iter()
-                        .map(|(name, ty)| (name.clone(), ty.clone()))
-                        .chain(
-                            policy_request
-                                .input
-                                .binding
-                                .claim_params
-                                .iter()
-                                .map(|(name, claim)| (name.clone(), claim.ty.clone())),
-                        ),
-                );
-                (
-                    GraphBuilder::binding_source(shape.clone(), descriptor),
-                    policy_request
-                        .input
-                        .binding
-                        .param_types
-                        .keys()
-                        .chain(policy_request.input.binding.claim_params.keys())
-                        .cloned()
-                        .collect::<BTreeSet<_>>(),
-                )
-            });
-        let authorized = match self.policy_authorization_row_id_graph(policy_request) {
-            Ok(authorized) => authorized,
-            Err(Error::QueryCapability(err)) if err.contains("PolicyProofCycle") => {
-                return Err(Error::QueryCapability(err));
-            }
-            Err(Error::QueryCapability(_err)) => PolicyAuthorizationGraph {
-                graph: empty_authorized_row_id_graph(),
-                route_fields: BTreeSet::new(),
-            },
-            Err(err) => return Err(err),
-        };
-        // Authorization is existential per protected row and binding route:
-        // multiple policy branches or multiple qualifying grant rows are
-        // alternative proofs, not additional copies of the application row.
-        // Collapse those proofs before the protected source reaches ordinary
-        // relational operators (especially finite TopBy windows). Route
-        // fields are part of the key because one prepared authorization graph
-        // can serve several independently routed bindings.
-        let mut authorization_keys = vec!["row_uuid".to_owned()];
-        authorization_keys.extend(authorized.route_fields.iter().cloned());
-        let authorized_graph = GraphBuilder::arg_max_by(
-            authorized.graph,
-            authorization_keys.clone(),
-            authorization_keys,
-        );
-        let (base, binding_route_fields) =
-            match binding_routes {
-                Some((binding, route_fields)) => (
-                    GraphBuilder::join(
-                        base,
-                        binding,
-                        std::iter::empty::<String>(),
-                        std::iter::empty::<String>(),
-                    )
-                    .project_fields(
-                        output_fields
-                            .iter()
-                            .map(|field| ProjectField::renamed(left_field(field), field.clone()))
-                            .chain(route_fields.iter().map(|field| {
-                                ProjectField::renamed(right_field(field), field.clone())
-                            }))
-                            .collect::<Vec<_>>(),
-                    ),
-                    route_fields,
-                ),
-                None => (base, BTreeSet::new()),
-            };
-        let mut join_keys = vec!["row_uuid".to_owned()];
-        join_keys.extend(authorized.route_fields.iter().cloned());
-        if authorized.route_fields.is_empty() {
-            let mut fields = output_fields
-                .iter()
-                .map(|field| ProjectField::renamed(left_field(&field), field.clone()))
-                .collect::<Vec<_>>();
-            fields.extend(
-                binding_route_fields
-                    .iter()
-                    .map(|field| ProjectField::renamed(left_field(field), field.clone())),
-            );
-            return Ok(PolicyAuthorizationGraph {
-                graph: GraphBuilder::join(base, authorized_graph, join_keys.clone(), join_keys)
-                    .project_fields(fields),
-                route_fields: binding_route_fields,
-            });
-        }
-        let mut fields = output_fields
-            .iter()
-            .map(|field| ProjectField::renamed(left_field(&field), field.clone()))
-            .collect::<Vec<_>>();
-        fields.extend(
-            authorized
-                .route_fields
-                .iter()
-                .map(|field| ProjectField::renamed(right_field(field), field.clone())),
-        );
-        fields.extend(
-            binding_route_fields
-                .iter()
-                .filter(|field| !authorized.route_fields.contains(*field))
-                .map(|field| ProjectField::renamed(left_field(field), field.clone())),
-        );
-        Ok(PolicyAuthorizationGraph {
-            graph: GraphBuilder::join(base, authorized_graph, join_keys.clone(), join_keys)
-                .project_fields(fields),
-            route_fields: binding_route_fields,
-        })
-    }
-
-    fn table_read_policy_authorization_request(
-        &mut self,
-        policy_schema_version: SchemaVersionId,
-        table_name: &str,
-        identity: AuthorId,
-        param_binding_mode: ParamBindingMode,
-        tier: DurabilityTier,
-        binding_source_shape: Option<String>,
-        binding_user_params: BTreeMap<String, ColumnType>,
-        binding_claim_params: BTreeMap<String, ProgramClaimParam>,
-    ) -> Result<QueryProgramRequest, Error> {
-        self.table_read_policy_authorization_request_with_root_visibility(
-            policy_schema_version,
-            table_name,
-            identity,
-            param_binding_mode,
-            tier,
-            binding_source_shape,
-            binding_user_params,
-            binding_claim_params,
-            false,
-        )
-    }
-
-    fn table_read_policy_authorization_request_at(
-        &self,
-        policy_schema_version: SchemaVersionId,
-        table_name: &str,
-        identity: AuthorId,
-        param_binding_mode: ParamBindingMode,
-        position: GlobalSeq,
-        binding_source_shape: Option<String>,
-        binding_user_params: BTreeMap<String, ColumnType>,
-        binding_claim_params: BTreeMap<String, ProgramClaimParam>,
-    ) -> Result<QueryProgramRequest, Error> {
-        let policy_schema = if policy_schema_version == self.catalogue.current_schema_version_id {
-            &self.catalogue.schema
-        } else {
-            &self
-                .catalogue
-                .catalogue_schemas
-                .get(&policy_schema_version)
-                .ok_or(Error::InvalidStoredValue(
-                    "policy schema version is unknown",
-                ))?
-                .schema
-        };
-        let table = policy_schema
-            .tables
-            .iter()
-            .find(|candidate| candidate.name == table_name)
-            .ok_or_else(|| Error::TableNotFound(table_name.to_owned()))?;
-        let query = authorization_query_from_read_policy(table);
-        if !query.includes.is_empty() {
-            return Err(Error::InvalidStoredValue(
-                "historical policy source filters do not support include policies",
-            ));
-        }
-        let policy_shape = query.validate(policy_schema)?;
-        let policy_binding = policy_shape.bind(BTreeMap::new())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            policy_schema,
-            param_binding_mode,
-        )?;
-        if !policy_shape.params().is_empty() {
-            return Err(Error::QueryCapability(
-                "historical policy source filters with runtime parameters must lower through query-engine binding sources"
-                    .to_owned(),
-            ));
-        }
-        let binding = policy_shape.bind(BTreeMap::new())?;
-        let mut input_shape = self.normalized_row_set_shape(&policy_shape, &binding)?;
-        let mut claim_params = binding_claim_params;
-        claim_params.extend(binding_claim_params_for_shape(
-            &input_shape,
-            policy_shape.params(),
-        ));
-        collect_reachable_seed_claim_params(
-            policy_schema,
-            policy_shape.query(),
-            &mut claim_params,
-        )?;
-        let binding_source_shape = binding_source_shape.clone().or_else(|| {
-            authorization_binding_source_shape(&policy_shape, &binding_user_params, &claim_params)
-        });
-        if let Some(source_shape) = binding_source_shape.clone() {
-            retarget_binding_value_sources(&mut input_shape, &source_shape);
-        }
-        let policy = match self.query_program_policy_context(identity) {
-            PolicyContext::Identity {
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            } => PolicyContext::AuthorizationSubplan {
-                protected_source: root_source_id(policy_shape.query().table.as_str()),
-                role: PolicyDecisionRole::Read,
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            },
-            other => other,
-        };
-        let input = RowSetProgramInput {
-            binding: self.program_binding_for_shape_and_policy(
-                &policy_shape,
-                &binding,
-                binding_source_shape,
-                binding_user_params,
-                claim_params,
-                &policy,
-            )?,
-            shape: input_shape,
-        };
-        Ok(QueryProgramRequest {
-            authorization_mode: QueryAuthorizationMode::TrustedServing,
-            reads: historical_query_read_set(&input.shape, policy_schema_version, position),
-            policy,
-            input,
-            output: current_query_output_request(
-                CurrentQueryProgramOutput::AuthorizedRows,
-                policy_shape.query(),
-            ),
-        })
-    }
-
-    fn table_read_policy_authorization_request_for_include_deleted(
-        &mut self,
-        policy_schema_version: SchemaVersionId,
-        table_name: &str,
-        identity: AuthorId,
-        tier: DurabilityTier,
-        binding_source_shape: Option<String>,
-        binding_user_params: BTreeMap<String, ColumnType>,
-        binding_claim_params: BTreeMap<String, ProgramClaimParam>,
-    ) -> Result<QueryProgramRequest, Error> {
-        self.table_read_policy_authorization_request_with_root_visibility(
-            policy_schema_version,
-            table_name,
-            identity,
-            ParamBindingMode::InlineAllReachableSeeds,
-            tier,
-            binding_source_shape,
-            binding_user_params,
-            binding_claim_params,
-            true,
-        )
-    }
-
-    fn table_read_policy_authorization_request_with_root_visibility(
-        &mut self,
-        policy_schema_version: SchemaVersionId,
-        table_name: &str,
-        identity: AuthorId,
-        param_binding_mode: ParamBindingMode,
-        tier: DurabilityTier,
-        binding_source_shape: Option<String>,
-        binding_user_params: BTreeMap<String, ColumnType>,
-        binding_claim_params: BTreeMap<String, ProgramClaimParam>,
-        include_deleted_root: bool,
-    ) -> Result<QueryProgramRequest, Error> {
-        let cache_key = ReadPolicyAuthorizationRequestCacheKey {
-            policy_schema_version,
-            table_name: table_name.to_owned(),
-            identity,
-            param_binding_mode: param_binding_mode.cache_key(),
-            tier,
-            binding_source_shape: binding_source_shape.clone(),
-            binding_user_params: binding_user_params_cache_key(&binding_user_params),
-            binding_claim_params: binding_claim_params_cache_key(&binding_claim_params),
-            include_deleted_root,
-        };
-        if let Some(request) = self
-            .query
-            .read_policy_authorization_request_cache
-            .get(&cache_key)
-        {
-            return Ok(request.clone());
-        }
-        let policy_schema = if policy_schema_version == self.catalogue.current_schema_version_id {
-            &self.catalogue.schema
-        } else {
-            &self
-                .catalogue
-                .catalogue_schemas
-                .get(&policy_schema_version)
-                .ok_or(Error::InvalidStoredValue(
-                    "policy schema version is unknown",
-                ))?
-                .schema
-        };
-        let table = policy_schema
-            .tables
-            .iter()
-            .find(|candidate| candidate.name == table_name)
-            .ok_or_else(|| Error::TableNotFound(table_name.to_owned()))?;
-        let policy = match self.query_program_policy_context(identity) {
-            PolicyContext::Identity {
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            } => PolicyContext::AuthorizationSubplan {
-                protected_source: root_source_id(table_name),
-                role: PolicyDecisionRole::Read,
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            },
-            other => other,
-        };
-        let mut query = authorization_query_from_read_policy(table);
-        let mut policy_binding_values = BTreeMap::new();
-        if matches!(param_binding_mode, ParamBindingMode::RetainAllParams)
-            && let PolicyContext::AuthorizationSubplan { claims, .. } = &policy
-        {
-            bind_scope_claim_operands(&mut query, claims, &mut policy_binding_values);
-        }
-        if !query.includes.is_empty() {
-            return Err(Error::InvalidStoredValue(
-                "maintained subscription view policy slice does not support include policies",
-            ));
-        }
-        let declared_claim_params = disambiguate_policy_claim_params_with_outer_slots(
-            &mut query,
-            policy_schema,
-            &mut policy_binding_values,
-            &binding_claim_params,
-        )?;
-        let policy_shape = query.validate(policy_schema)?;
-        coerce_binding_values_for_shape(&policy_shape, &mut policy_binding_values);
-        let policy_binding = policy_shape.bind(policy_binding_values.clone())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            policy_schema,
-            param_binding_mode,
-        )?;
-        if policy_shape
-            .params()
-            .keys()
-            .any(|name| !policy_binding_values.contains_key(name))
-        {
-            return Err(Error::QueryCapability(
-                "maintained policy source filters with runtime parameters must lower through query-engine binding sources"
-                    .to_owned(),
-            ));
-        }
-        let binding = policy_shape.bind(policy_binding_values)?;
-        let mut input_shape = if include_deleted_root {
-            self.normalized_include_deleted_row_set_shape(&policy_shape, &binding)?
-        } else {
-            self.normalized_row_set_shape(&policy_shape, &binding)?
-        };
-        let mut claim_params = binding_claim_params;
-        claim_params.extend(binding_claim_params_for_shape(
-            &input_shape,
-            policy_shape.params(),
-        ));
-        claim_params.extend(declared_claim_params);
-        collect_reachable_seed_claim_params(
-            policy_schema,
-            policy_shape.query(),
-            &mut claim_params,
-        )?;
-        for (name, claim) in &mut claim_params {
-            if let Some(ty) = policy_shape.params().get(name) {
-                claim.ty = ty.clone();
-            }
-        }
-        let binding_source_shape = binding_source_shape.clone().or_else(|| {
-            authorization_binding_source_shape(&policy_shape, &binding_user_params, &claim_params)
-        });
-        if let Some(source_shape) = binding_source_shape.clone() {
-            retarget_binding_value_sources(&mut input_shape, &source_shape);
-        }
-        let input = RowSetProgramInput {
-            binding: self.program_binding_for_shape_and_policy(
-                &policy_shape,
-                &binding,
-                binding_source_shape,
-                binding_user_params,
-                claim_params,
-                &policy,
-            )?,
-            shape: input_shape,
-        };
-        let request = QueryProgramRequest {
-            authorization_mode: QueryAuthorizationMode::TrustedServing,
-            reads: current_query_read_set(
-                &input.shape,
-                policy_schema_version,
-                policy_schema_version,
-                tier,
-                None,
-            ),
-            policy,
-            input,
-            output: current_query_output_request(
-                CurrentQueryProgramOutput::AuthorizedRows,
-                policy_shape.query(),
-            ),
-        };
-        self.query
-            .read_policy_authorization_request_cache
-            .insert(cache_key, request.clone());
-        Ok(request)
-    }
-
-    fn branch_table_read_policy_authorization_request(
-        &self,
-        branch_id: BranchId,
-        table: &TableSchema,
-        identity: AuthorId,
-        binding_source_shape: Option<String>,
-        binding_user_params: BTreeMap<String, ColumnType>,
-        binding_claim_params: BTreeMap<String, ProgramClaimParam>,
-    ) -> Result<QueryProgramRequest, Error> {
-        let query = authorization_query_from_read_policy(table);
-        if !query.includes.is_empty() {
-            return Err(Error::InvalidStoredValue(
-                "branch policy source filters do not support include policies",
-            ));
-        }
-        let policy_shape = query.validate(&self.catalogue.schema)?;
-        let policy_binding = policy_shape.bind(BTreeMap::new())?;
-        let policy_shape = bind_query_params_with_mode(
-            &policy_shape,
-            &policy_binding,
-            &self.catalogue.schema,
-            ParamBindingMode::InlineAllReachableSeeds,
-        )?;
-        if !policy_shape.params().is_empty() {
-            return Err(Error::QueryCapability(
-                "branch policy source filters with runtime parameters must lower through query-engine binding sources"
-                    .to_owned(),
-            ));
-        }
-        let binding = policy_shape.bind(BTreeMap::new())?;
-        let mut input_shape = self.normalized_row_set_shape(&policy_shape, &binding)?;
-        let mut claim_params = binding_claim_params;
-        claim_params.extend(binding_claim_params_for_shape(
-            &input_shape,
-            policy_shape.params(),
-        ));
-        collect_reachable_seed_claim_params(
-            &self.catalogue.schema,
-            policy_shape.query(),
-            &mut claim_params,
-        )?;
-        let binding_source_shape = binding_source_shape.clone().or_else(|| {
-            authorization_binding_source_shape(&policy_shape, &binding_user_params, &claim_params)
-        });
-        if let Some(source_shape) = binding_source_shape.clone() {
-            retarget_binding_value_sources(&mut input_shape, &source_shape);
-        }
-        let policy = match self.query_program_policy_context(identity) {
-            PolicyContext::Identity {
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            } => PolicyContext::AuthorizationSubplan {
-                protected_source: root_source_id(policy_shape.query().table.as_str()),
-                role: PolicyDecisionRole::Read,
-                mode,
-                permission_subject,
-                claims,
-                attribution,
-            },
-            other => other,
-        };
-        let input = RowSetProgramInput {
-            binding: self.program_binding_for_shape_and_policy(
-                &policy_shape,
-                &binding,
-                binding_source_shape,
-                binding_user_params,
-                claim_params,
-                &policy,
-            )?,
-            shape: input_shape,
-        };
-        Ok(QueryProgramRequest {
-            authorization_mode: QueryAuthorizationMode::TrustedServing,
-            reads: branch_query_read_set(
-                &input.shape,
-                policy_shape.schema_version(),
-                DurabilityTier::Local,
-                branch_id,
-            ),
-            policy,
-            input,
-            output: current_query_output_request(
-                CurrentQueryProgramOutput::AuthorizedRows,
-                policy_shape.query(),
-            ),
-        })
-    }
-
-    fn maintained_view_content_current_with_version(
-        &self,
-        table: &TableSchema,
-        tier: DurabilityTier,
-    ) -> Result<GraphBuilder, Error> {
-        let schema_version = self.catalogue.current_schema_version_id;
-        self.maintained_view_content_current_with_version_in_schema(table, tier, schema_version)
-    }
-
-    fn maintained_view_content_current_with_version_in_schema(
-        &self,
-        table: &TableSchema,
-        tier: DurabilityTier,
-        schema_version: SchemaVersionId,
-    ) -> Result<GraphBuilder, Error> {
-        let table_id = self.physical_table_id_for_schema(schema_version, &table.name)?;
-        let content_fields = global_current_storage_fields(table, true, true);
-        let global_content = self
-            .physical_current_source_graph(
-                schema_version,
-                &table.name,
-                PhysicalCurrentClass::Global,
-            )?
-            .project(content_fields.clone());
-        let global_deletion =
-            GraphBuilder::table(physical_register_global_current_table_name(table_id))
-                .project_fields(register_storage_fields_for_query_engine(""));
-
-        let (content, deletion) = if tier == DurabilityTier::Global {
-            (global_content, global_deletion)
-        } else {
-            let ahead_content = self.physical_current_source_graph(
-                schema_version,
-                &table.name,
-                PhysicalCurrentClass::Ahead,
-            )?;
-            let ahead_content = if tier == DurabilityTier::Edge {
-                edge_visible_ahead_current_source_graph(ahead_content, content_fields.clone())
-            } else {
-                ahead_content.project(content_fields.clone())
-            };
-            let ahead_deletion =
-                GraphBuilder::table(physical_register_ahead_current_table_name(table_id));
-            let ahead_deletion = if tier == DurabilityTier::Edge {
-                edge_visible_ahead_current_source_graph(
-                    ahead_deletion,
-                    register_storage_field_names(),
-                )
-            } else {
-                ahead_deletion.project_fields(register_storage_fields_for_query_engine(""))
-            };
-            (
-                GraphBuilder::arg_max_by(
-                    GraphBuilder::union([global_content, ahead_content]),
-                    ["row_uuid"],
-                    ["tx_time", "tx_node_id"],
-                )
-                .project(content_fields),
-                GraphBuilder::arg_max_by(
-                    GraphBuilder::union([global_deletion, ahead_deletion]),
-                    ["row_uuid"],
-                    ["tx_time", "tx_node_id"],
-                )
-                .project_fields(register_storage_fields_for_query_engine("")),
-            )
-        };
-        let deleted = deletion
-            .filter(PredicateExpr::eq("_deletion", Value::EnumTag(0)))
-            .project(["row_uuid"]);
-        Ok(GraphBuilder::anti_join(
-            content,
-            deleted,
-            ["row_uuid"],
-            ["row_uuid"],
-        ))
-    }
-
-    #[cfg(test)]
-    pub(crate) fn test_content_current_with_version(
-        &mut self,
-        table: &TableSchema,
-        tier: DurabilityTier,
-    ) -> Result<groove::ivm::RecordDeltas, Error> {
-        let graph = self.maintained_view_content_current_with_version(table, tier)?;
-        self.database.query_graph(graph).map_err(Error::Groove)
-    }
-
-    /// Pick the policy-owning schema for one authorization-support operation.
-    ///
-    /// A newer read-policy revision must not shadow an older write-policy
-    /// revision (or vice versa): the terminal proof must hydrate precisely the
-    /// policy clause that admission will evaluate.
-    fn authorization_scope_policy_schema_for_action(
-        &self,
-        table: &str,
-        operation: AuthorizationScopeOperation,
-    ) -> SchemaVersionId {
-        let write_schema = self.catalogue.current_write_schema.schema;
-        let has_operation_policy = self
-            .table_in_schema(table, write_schema)
-            .is_ok_and(|table| match operation {
-                AuthorizationScopeOperation::Read => {
-                    table.read_policy.is_some() || access_edge_parent_reference(&table).is_some()
-                }
-                AuthorizationScopeOperation::Insert
-                | AuthorizationScopeOperation::Update
-                | AuthorizationScopeOperation::Delete => {
-                    !authorization_policy_queries(&table, operation).is_empty()
-                }
-            });
-        if has_operation_policy {
-            write_schema
-        } else {
-            self.catalogue.current_schema_version_id
-        }
-    }
-
-    /// Compile the exact policy clauses needed for one non-mutating operation.
-    /// This is intentionally separate from the legacy table-wide write scope:
-    /// callers are not switched until the receipt transport is negotiated.
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub(crate) fn authorization_support_scope(
-        &self,
-        writer: AuthorId,
-        action: &PermissionAdviceAction,
-    ) -> Result<AuthorizationSupportScope, Error> {
-        let (operation, table_name) = authorization_scope_action(action);
-        // `PermissionAdviceAction` is reconstructed after policy projection,
-        // so its table belongs to the policy-owning schema, not necessarily
-        // this node's base API schema. In particular, a rename lens can turn
-        // an authored `users` version into a `people` policy action while an
-        // old-schema subscriber remains live. Resolve both the policy and its
-        // support-query schema from that policy view.
-        let policy_schema_version =
-            self.authorization_scope_policy_schema_for_action(table_name, operation);
-        let policy_schema = self
-            .catalogue
-            .catalogue_schemas
-            .get(&policy_schema_version)
-            .ok_or(Error::InvalidStoredValue(
-                "authorization policy schema is unknown",
-            ))?;
-        let policies = authorization_policy_queries(
-            &self.table_in_schema(table_name, policy_schema_version)?,
-            operation,
-        );
-        let claims = self.session_claims.get(&writer);
-        let mut claim_values = default_permission_scope_claim_values(writer);
-        if let Some(claims) = claims {
-            claim_values.extend(claims.clone());
-        }
-        // Authorization support is authority-current: historic/branch views
-        // and weaker durability tiers cannot vouch for the authoritative edge.
-        let options = RegisterShapeOptions::default();
-        let subscriptions = policies
-            .iter()
-            .map(|policy| {
-                compile_permission_scope_policy(
-                    policy.clone(),
-                    claims,
-                    &claim_values,
-                    &policy_schema.schema,
-                )
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        let policy_bytes = postcard::to_allocvec(&(operation, &policies))
-            .map_err(|_| Error::InvalidStoredValue("authorization policy serialization failed"))?;
-        let claim_bytes = postcard::to_allocvec(&claim_values)
-            .map_err(|_| Error::InvalidStoredValue("authorization claims serialization failed"))?;
-        let support_bytes = postcard::to_allocvec(&(
-            operation,
-            &options,
-            subscriptions
-                .iter()
-                .map(|(shape, binding)| (shape.shape_id(), binding.binding_id()))
-                .collect::<Vec<_>>(),
-        ))
-        .map_err(|_| Error::InvalidStoredValue("authorization scope serialization failed"))?;
-        let action_bytes = postcard::to_allocvec(action).map_err(|_| {
-            Error::InvalidStoredValue("authorization operation serialization failed")
-        })?;
-        let operation_key =
-            authorization_operation_key(operation, table_name, action, action_bytes);
-        Ok(AuthorizationSupportScope {
-            key: AuthorizationSupportScopeKey {
-                support_shape_digest: *blake3::hash(&support_bytes).as_bytes(),
-                subject: writer,
-                claims_digest: *blake3::hash(&claim_bytes).as_bytes(),
-                policy_digest: *blake3::hash(&policy_bytes).as_bytes(),
-            },
-            operation: operation_key,
-            options,
-            subscriptions,
-        })
-    }
-}
-
-fn empty_policy_filtered_current_source_graph(
-    base: GraphBuilder,
-    output_fields: &[String],
-) -> PolicyAuthorizationGraph {
-    let keys = ["row_uuid".to_owned()];
-    PolicyAuthorizationGraph {
-        graph: GraphBuilder::join(base, empty_authorized_row_id_graph(), keys.clone(), keys)
-            .project_fields(
-                output_fields
-                    .iter()
-                    .map(|field| ProjectField::renamed(left_field(field), field.clone()))
-                    .collect::<Vec<_>>(),
-            ),
-        route_fields: BTreeSet::new(),
-    }
 }
 
 /// Wrap a compiler aggregate record in the minimal [`CurrentRow`] envelope.
@@ -3970,126 +2652,7 @@ fn aggregate_current_row_from_record(
     ))
 }
 
-#[cfg_attr(not(test), allow(dead_code))]
-fn compile_permission_scope_policy(
-    mut query: JazzQuery,
-    claims: Option<&BTreeMap<String, Value>>,
-    claim_values: &BTreeMap<String, Value>,
-    schema: &JazzSchema,
-) -> Result<(ValidatedQuery, Binding), Error> {
-    query.filters = query
-        .filters
-        .into_iter()
-        .map(|p| rewrite_claim_predicate_for_binding(p, claims))
-        .collect();
-    query.joins = query
-        .joins
-        .into_iter()
-        .map(|j| rewrite_claim_join_for_binding(j, claims))
-        .collect();
-    query.reachable = query
-        .reachable
-        .into_iter()
-        .map(|mut reachable| {
-            reachable.access_filters = reachable
-                .access_filters
-                .into_iter()
-                .map(|p| rewrite_claim_predicate_for_binding(p, claims))
-                .collect();
-            reachable.edge_filters = reachable
-                .edge_filters
-                .into_iter()
-                .map(|p| rewrite_claim_predicate_for_binding(p, claims))
-                .collect();
-            if let Some(seed) = &mut reachable.seed {
-                seed.filters = std::mem::take(&mut seed.filters)
-                    .into_iter()
-                    .map(|p| rewrite_claim_predicate_for_binding(p, claims))
-                    .collect();
-            }
-            reachable
-        })
-        .collect();
-    let mut values = BTreeMap::new();
-    bind_scope_claim_operands(&mut query, claim_values, &mut values);
-    let shape = query.validate(schema)?;
-    coerce_binding_values_for_shape(&shape, &mut values);
-    let binding = shape.bind(values)?;
-    Ok((shape, binding))
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-fn authorization_scope_action(
-    action: &PermissionAdviceAction,
-) -> (AuthorizationScopeOperation, &str) {
-    match action {
-        PermissionAdviceAction::Read { table, .. } => (AuthorizationScopeOperation::Read, table),
-        PermissionAdviceAction::Insert { table, .. } => {
-            (AuthorizationScopeOperation::Insert, table)
-        }
-        PermissionAdviceAction::Update { table, .. } => {
-            (AuthorizationScopeOperation::Update, table)
-        }
-        PermissionAdviceAction::Delete { table, .. } => {
-            (AuthorizationScopeOperation::Delete, table)
-        }
-    }
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-fn authorization_policy_queries(
-    table: &crate::schema::TableSchema,
-    operation: AuthorizationScopeOperation,
-) -> Vec<JazzQuery> {
-    match operation {
-        AuthorizationScopeOperation::Read
-            if table.read_policy.is_none() && access_edge_parent_reference(table).is_none() =>
-        {
-            Vec::new()
-        }
-        AuthorizationScopeOperation::Read => vec![authorization_query_from_read_policy(table)],
-        AuthorizationScopeOperation::Insert => table
-            .write_policies
-            .insert_check
-            .clone()
-            .into_iter()
-            .collect(),
-        AuthorizationScopeOperation::Update => [
-            table.write_policies.update_using.clone(),
-            table.write_policies.update_check.clone(),
-        ]
-        .into_iter()
-        .flatten()
-        .collect(),
-        AuthorizationScopeOperation::Delete => table
-            .write_policies
-            .delete_using
-            .clone()
-            .into_iter()
-            .collect(),
-    }
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-fn authorization_operation_key(
-    operation: AuthorizationScopeOperation,
-    table: &str,
-    action: &PermissionAdviceAction,
-    action_bytes: Vec<u8>,
-) -> AuthorizationOperationKey {
-    let row = match action {
-        PermissionAdviceAction::Read { row, .. }
-        | PermissionAdviceAction::Update { row, .. }
-        | PermissionAdviceAction::Delete { row, .. } => Some(*row),
-        PermissionAdviceAction::Insert { .. } => None,
-    };
-    AuthorizationOperationKey {
-        operation,
-        table: table.to_owned(),
-        row,
-        candidate_digest: *blake3::hash(&action_bytes).as_bytes(),
-    }
-}
+mod authorization;
 
 #[cfg(test)]
 mod authorization_scope_compiler_tests {
@@ -4562,1113 +3125,9 @@ where
     }
 }
 
-fn authorization_query_from_read_policy(table: &TableSchema) -> JazzQuery {
-    let Some(policy) = &table.read_policy else {
-        return crate::query::Query::from(table.name.as_str());
-    };
-    let mut query = crate::query::Query::from(table.name.as_str());
-    query.filters = policy.filters.clone();
-    query.joins = policy.joins.clone();
-    query.reachable = policy.reachable.clone();
-    query.inherits = policy.inherits.clone();
-    query.includes = policy.includes.clone();
-    query.policy_branches = policy.policy_branches.clone();
-    if let Some(parent_column) = access_edge_parent_reference(table) {
-        query.policy_branches.push(crate::query::PolicyBranch {
-            filters: Vec::new(),
-            joins: Vec::new(),
-            reachable: Vec::new(),
-            inherits: vec![crate::query::InheritsVia {
-                parent_column,
-                operation: crate::query::InheritsOperation::Select,
-                max_depth: None,
-            }],
-        });
-    }
-    query
-}
+mod bindings;
 
-fn access_edge_parent_reference(table: &TableSchema) -> Option<String> {
-    if !table.name.ends_with("_access_edges") && table.name != "team_access_edges" {
-        return None;
-    }
-    table
-        .references
-        .contains_key("resource_id")
-        .then(|| "resource_id".to_owned())
-}
-
-fn rewrite_claim_join_for_binding(
-    join: JoinVia,
-    claims: Option<&BTreeMap<String, Value>>,
-) -> JoinVia {
-    JoinVia {
-        table: join.table,
-        on_column: join.on_column,
-        target: join.target,
-        source_column: join.source_column,
-        source_lookup: join.source_lookup,
-        correlated_filters: join.correlated_filters,
-        filters: join
-            .filters
-            .into_iter()
-            .map(|predicate| rewrite_claim_predicate_for_binding(predicate, claims))
-            .collect(),
-        nested_joins: join
-            .nested_joins
-            .into_iter()
-            .map(|join| rewrite_claim_join_for_binding(join, claims))
-            .collect(),
-    }
-}
-
-fn rewrite_claim_predicate_for_binding(
-    predicate: Predicate,
-    claims: Option<&BTreeMap<String, Value>>,
-) -> Predicate {
-    match predicate {
-        Predicate::All(predicates) => Predicate::All(
-            predicates
-                .into_iter()
-                .map(|predicate| rewrite_claim_predicate_for_binding(predicate, claims))
-                .collect(),
-        ),
-        Predicate::Any(predicates) => Predicate::Any(
-            predicates
-                .into_iter()
-                .map(|predicate| rewrite_claim_predicate_for_binding(predicate, claims))
-                .collect(),
-        ),
-        Predicate::Not(predicate) if predicate_contains_unbound_claim(&predicate, claims) => {
-            false_predicate()
-        }
-        Predicate::Not(predicate) => Predicate::Not(Box::new(rewrite_claim_predicate_for_binding(
-            *predicate, claims,
-        ))),
-        Predicate::Eq(left, right) if operands_contain_unbound_claim([&left, &right], claims) => {
-            false_predicate()
-        }
-        Predicate::Eq(left, right) => Predicate::Eq(left, right),
-        Predicate::Ne(left, right) if operands_contain_unbound_claim([&left, &right], claims) => {
-            false_predicate()
-        }
-        Predicate::Ne(left, right) => Predicate::Ne(left, right),
-        Predicate::In(left, values)
-            if operands_contain_unbound_claim(
-                std::iter::once(&left)
-                    .chain(values.iter())
-                    .collect::<Vec<_>>(),
-                claims,
-            ) =>
-        {
-            false_predicate()
-        }
-        Predicate::In(left, values) => Predicate::In(left, values),
-        Predicate::Gt(_, _) | Predicate::Gte(_, _) | Predicate::Lt(_, _) | Predicate::Lte(_, _) => {
-            false_predicate()
-        }
-        Predicate::Contains(left, right)
-            if operands_contain_unbound_claim([&left, &right], claims) =>
-        {
-            false_predicate()
-        }
-        Predicate::Contains(left, right) => Predicate::Contains(left, right),
-        Predicate::EnumMatch {
-            column,
-            case,
-            payload,
-        } => Predicate::EnumMatch {
-            column,
-            case,
-            payload: Box::new(rewrite_claim_predicate_for_binding(*payload, claims)),
-        },
-        Predicate::IsNull(_) => false_predicate(),
-    }
-}
-
-fn default_permission_scope_claim_values(writer: AuthorId) -> BTreeMap<String, Value> {
-    default_policy_claim_values(writer)
-}
-
-fn default_policy_claim_values(writer: AuthorId) -> BTreeMap<String, Value> {
-    // Alpha-compat built-ins live at the node admission/query boundary, not in
-    // the compiler: lowering receives ordinary claim values plus spec `sub`.
-    BUILTIN_POLICY_CLAIMS
-        .iter()
-        .map(|name| {
-            let value = match *name {
-                "sub" => Value::Uuid(writer.0),
-                "user_id" => Value::String(writer.0.to_string()),
-                "isAdmin" => Value::Bool(false),
-                _ => unreachable!("unknown built-in policy claim"),
-            };
-            ((*name).to_owned(), value)
-        })
-        .collect()
-}
-
-const BUILTIN_POLICY_CLAIMS: &[&str] = &["sub", "user_id", "isAdmin"];
-
-fn is_builtin_policy_claim(name: &str) -> bool {
-    BUILTIN_POLICY_CLAIMS.contains(&name)
-}
-
-fn bind_scope_claim_operands(
-    query: &mut JazzQuery,
-    claim_values: &BTreeMap<String, Value>,
-    binding_values: &mut BTreeMap<String, Value>,
-) {
-    for predicate in &mut query.filters {
-        bind_scope_claim_predicate(predicate, claim_values, binding_values);
-    }
-    for join in &mut query.joins {
-        bind_scope_claim_join(join, claim_values, binding_values);
-    }
-    for reachable in &mut query.reachable {
-        for predicate in &mut reachable.access_filters {
-            bind_scope_claim_predicate(predicate, claim_values, binding_values);
-        }
-        for predicate in &mut reachable.edge_filters {
-            bind_scope_claim_predicate(predicate, claim_values, binding_values);
-        }
-        if let Some(seed) = &mut reachable.seed {
-            for predicate in &mut seed.filters {
-                bind_scope_claim_predicate(predicate, claim_values, binding_values);
-            }
-        }
-    }
-    for branch in &mut query.policy_branches {
-        for predicate in &mut branch.filters {
-            bind_scope_claim_predicate(predicate, claim_values, binding_values);
-        }
-        for join in &mut branch.joins {
-            bind_scope_claim_join(join, claim_values, binding_values);
-        }
-        for reachable in &mut branch.reachable {
-            for predicate in &mut reachable.access_filters {
-                bind_scope_claim_predicate(predicate, claim_values, binding_values);
-            }
-            for predicate in &mut reachable.edge_filters {
-                bind_scope_claim_predicate(predicate, claim_values, binding_values);
-            }
-            if let Some(seed) = &mut reachable.seed {
-                for predicate in &mut seed.filters {
-                    bind_scope_claim_predicate(predicate, claim_values, binding_values);
-                }
-            }
-        }
-    }
-}
-
-fn bind_scope_claim_join(
-    join: &mut JoinVia,
-    claim_values: &BTreeMap<String, Value>,
-    binding_values: &mut BTreeMap<String, Value>,
-) {
-    for predicate in &mut join.filters {
-        bind_scope_claim_predicate(predicate, claim_values, binding_values);
-    }
-    for join in &mut join.nested_joins {
-        bind_scope_claim_join(join, claim_values, binding_values);
-    }
-}
-
-fn bind_scope_claim_predicate(
-    predicate: &mut Predicate,
-    claim_values: &BTreeMap<String, Value>,
-    binding_values: &mut BTreeMap<String, Value>,
-) {
-    match predicate {
-        Predicate::All(predicates) | Predicate::Any(predicates) => {
-            for predicate in predicates {
-                bind_scope_claim_predicate(predicate, claim_values, binding_values);
-            }
-        }
-        Predicate::Not(predicate) => {
-            bind_scope_claim_predicate(predicate, claim_values, binding_values);
-        }
-        Predicate::Eq(left, right)
-        | Predicate::Ne(left, right)
-        | Predicate::Gt(left, right)
-        | Predicate::Gte(left, right)
-        | Predicate::Lt(left, right)
-        | Predicate::Lte(left, right)
-        | Predicate::Contains(left, right) => {
-            bind_scope_claim_operand(left, claim_values, binding_values);
-            bind_scope_claim_operand(right, claim_values, binding_values);
-        }
-        Predicate::In(left, values) => {
-            bind_scope_claim_operand(left, claim_values, binding_values);
-            for value in values {
-                bind_scope_claim_operand(value, claim_values, binding_values);
-            }
-        }
-        Predicate::IsNull(operand) => {
-            bind_scope_claim_operand(operand, claim_values, binding_values);
-        }
-        Predicate::EnumMatch { payload, .. } => {
-            bind_scope_claim_predicate(payload, claim_values, binding_values);
-        }
-    }
-}
-
-fn bind_scope_claim_operand(
-    operand: &mut Operand,
-    claim_values: &BTreeMap<String, Value>,
-    binding_values: &mut BTreeMap<String, Value>,
-) {
-    let Operand::Claim(name) = operand else {
-        return;
-    };
-    let Some(value) = claim_values.get(name).cloned() else {
-        return;
-    };
-    let param = claim_param_field(&ClaimPath(vec![name.clone()]));
-    binding_values.insert(param.clone(), value);
-    *operand = Operand::Param(param);
-}
-
-fn disambiguate_policy_claim_params(
-    query: &mut JazzQuery,
-    schema: &JazzSchema,
-    binding_values: &mut BTreeMap<String, Value>,
-) -> Result<BTreeMap<String, ProgramClaimParam>, Error> {
-    disambiguate_policy_claim_params_with_outer_slots(
-        query,
-        schema,
-        binding_values,
-        &BTreeMap::new(),
-    )
-}
-
-/// Give a policy-local claim parameter a stable binding slot. A nested policy
-/// which is lowered under an already-prepared outer source must reuse that
-/// source's slot when its claim path and validated type are identical. Creating
-/// a fresh typed alias in that case changes the shared source descriptor after
-/// it was registered. Different validated types deliberately retain distinct
-/// aliases, so a claim cannot cross a type boundary through source reuse.
-fn disambiguate_policy_claim_params_with_outer_slots(
-    query: &mut JazzQuery,
-    schema: &JazzSchema,
-    binding_values: &mut BTreeMap<String, Value>,
-    outer_slots: &BTreeMap<String, ProgramClaimParam>,
-) -> Result<BTreeMap<String, ProgramClaimParam>, Error> {
-    let shape = query.validate(schema)?;
-    let mut aliases = BTreeMap::new();
-    let mut claims = BTreeMap::new();
-    for (name, ty) in shape.params() {
-        let Some(path) = claim_path_from_param_field(name) else {
-            continue;
-        };
-        let alias = outer_slots
-            .iter()
-            .find(|(_, slot)| slot.path == path && slot.ty == *ty)
-            .map(|(name, _)| name.clone())
-            .unwrap_or_else(|| typed_claim_param_alias(name, ty));
-        aliases.insert(name.clone(), alias.clone());
-        claims.insert(
-            alias,
-            ProgramClaimParam {
-                path,
-                ty: ty.clone(),
-            },
-        );
-    }
-    rename_query_params(query, &aliases);
-    for (name, alias) in aliases {
-        if let Some(value) = binding_values.remove(&name) {
-            binding_values.insert(alias, value);
-        }
-    }
-    Ok(claims)
-}
-
-fn typed_claim_param_alias(name: &str, ty: &ColumnType) -> String {
-    let ty = format!("{ty:?}");
-    format!("__jazz_claim_typed:{}:{ty}:{name}", ty.len())
-}
-
-fn rename_query_params(query: &mut JazzQuery, aliases: &BTreeMap<String, String>) {
-    for predicate in &mut query.filters {
-        rename_predicate_params(predicate, aliases);
-    }
-    for join in &mut query.joins {
-        rename_join_params(join, aliases);
-    }
-    for reachable in &mut query.reachable {
-        rename_reachable_params(reachable, aliases);
-    }
-    for branch in &mut query.policy_branches {
-        for predicate in &mut branch.filters {
-            rename_predicate_params(predicate, aliases);
-        }
-        for join in &mut branch.joins {
-            rename_join_params(join, aliases);
-        }
-        for reachable in &mut branch.reachable {
-            rename_reachable_params(reachable, aliases);
-        }
-    }
-}
-
-fn rename_join_params(join: &mut JoinVia, aliases: &BTreeMap<String, String>) {
-    for predicate in &mut join.filters {
-        rename_predicate_params(predicate, aliases);
-    }
-    for join in &mut join.nested_joins {
-        rename_join_params(join, aliases);
-    }
-}
-
-fn rename_reachable_params(
-    reachable: &mut crate::query::ReachableVia,
-    aliases: &BTreeMap<String, String>,
-) {
-    rename_operand_param(&mut reachable.from, aliases);
-    for predicate in &mut reachable.access_filters {
-        rename_predicate_params(predicate, aliases);
-    }
-    for predicate in &mut reachable.edge_filters {
-        rename_predicate_params(predicate, aliases);
-    }
-    if let Some(seed) = &mut reachable.seed {
-        for predicate in &mut seed.filters {
-            rename_predicate_params(predicate, aliases);
-        }
-    }
-}
-
-fn rename_predicate_params(predicate: &mut Predicate, aliases: &BTreeMap<String, String>) {
-    match predicate {
-        Predicate::All(predicates) | Predicate::Any(predicates) => {
-            for predicate in predicates {
-                rename_predicate_params(predicate, aliases);
-            }
-        }
-        Predicate::Not(predicate) => rename_predicate_params(predicate, aliases),
-        Predicate::Eq(left, right)
-        | Predicate::Ne(left, right)
-        | Predicate::Gt(left, right)
-        | Predicate::Gte(left, right)
-        | Predicate::Lt(left, right)
-        | Predicate::Lte(left, right)
-        | Predicate::Contains(left, right) => {
-            rename_operand_param(left, aliases);
-            rename_operand_param(right, aliases);
-        }
-        Predicate::In(left, values) => {
-            rename_operand_param(left, aliases);
-            for value in values {
-                rename_operand_param(value, aliases);
-            }
-        }
-        Predicate::IsNull(operand) => rename_operand_param(operand, aliases),
-        Predicate::EnumMatch { payload, .. } => rename_predicate_params(payload, aliases),
-    }
-}
-
-fn rename_operand_param(operand: &mut Operand, aliases: &BTreeMap<String, String>) {
-    let Operand::Param(name) = operand else {
-        return;
-    };
-    if let Some(alias) = aliases.get(name) {
-        *name = alias.clone();
-    }
-}
-
-fn false_predicate() -> Predicate {
-    Predicate::Eq(
-        Operand::Literal(Value::Bool(true)),
-        Operand::Literal(Value::Bool(false)),
-    )
-}
-
-fn predicate_contains_unbound_claim(
-    predicate: &Predicate,
-    claims: Option<&BTreeMap<String, Value>>,
-) -> bool {
-    match predicate {
-        Predicate::All(predicates) | Predicate::Any(predicates) => predicates
-            .iter()
-            .any(|predicate| predicate_contains_unbound_claim(predicate, claims)),
-        Predicate::Not(predicate) => predicate_contains_unbound_claim(predicate, claims),
-        Predicate::Eq(left, right)
-        | Predicate::Ne(left, right)
-        | Predicate::Gt(left, right)
-        | Predicate::Gte(left, right)
-        | Predicate::Lt(left, right)
-        | Predicate::Lte(left, right)
-        | Predicate::Contains(left, right) => operands_contain_unbound_claim([left, right], claims),
-        Predicate::In(left, values) => {
-            operand_contains_unbound_claim(left, claims)
-                || values
-                    .iter()
-                    .any(|operand| operand_contains_unbound_claim(operand, claims))
-        }
-        Predicate::IsNull(operand) => operand_contains_unbound_claim(operand, claims),
-        Predicate::EnumMatch { payload, .. } => predicate_contains_unbound_claim(payload, claims),
-    }
-}
-
-fn operands_contain_unbound_claim<'a>(
-    operands: impl IntoIterator<Item = &'a Operand>,
-    claims: Option<&BTreeMap<String, Value>>,
-) -> bool {
-    operands
-        .into_iter()
-        .any(|operand| operand_contains_unbound_claim(operand, claims))
-}
-
-fn operand_contains_unbound_claim(
-    operand: &Operand,
-    claims: Option<&BTreeMap<String, Value>>,
-) -> bool {
-    matches!(operand, Operand::Claim(name) if !is_builtin_policy_claim(name) && !claims.is_some_and(|claims| claims.contains_key(name)))
-}
-
-#[derive(Clone, Copy)]
-pub(crate) enum ParamBindingMode {
-    InlineAllReachableSeeds,
-    RetainAllParams,
-}
-
-impl ParamBindingMode {
-    fn cache_key(self) -> ParamBindingModeCacheKey {
-        match self {
-            Self::InlineAllReachableSeeds => ParamBindingModeCacheKey::InlineAllReachableSeeds,
-            Self::RetainAllParams => ParamBindingModeCacheKey::RetainAllParams,
-        }
-    }
-}
-
-fn binding_user_params_cache_key(params: &BTreeMap<String, ColumnType>) -> String {
-    format!("{params:?}")
-}
-
-fn binding_claim_params_cache_key(params: &BTreeMap<String, ProgramClaimParam>) -> String {
-    format!("{params:?}")
-}
-
-fn bind_query_params_with_mode(
-    shape: &ValidatedQuery,
-    binding: &Binding,
-    schema: &JazzSchema,
-    mode: ParamBindingMode,
-) -> Result<ValidatedQuery, Error> {
-    let mut query = shape.query().clone();
-    let root_source = root_source_id(&query.table);
-    query.filters = query
-        .filters
-        .into_iter()
-        .map(|predicate| bind_query_predicate(predicate, binding, schema, &root_source, mode))
-        .collect::<Result<Vec<_>, _>>()?;
-    query.joins = query
-        .joins
-        .into_iter()
-        .map(|join| bind_join_filter_literals(join, binding, schema, mode))
-        .collect::<Result<Vec<_>, Error>>()?;
-    query.reachable = query
-        .reachable
-        .into_iter()
-        .map(|mut reachable| {
-            if should_inline_reachable_seed(&reachable.from, mode) {
-                reachable.from = bind_query_operand(reachable.from, binding, mode)?;
-            }
-            reachable.access_filters = reachable
-                .access_filters
-                .into_iter()
-                .map(|predicate| {
-                    bind_query_predicate(
-                        predicate,
-                        binding,
-                        schema,
-                        &bind_source_for_table(&reachable.access_table),
-                        mode,
-                    )
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-            reachable.edge_filters = reachable
-                .edge_filters
-                .into_iter()
-                .map(|predicate| {
-                    bind_query_predicate(
-                        predicate,
-                        binding,
-                        schema,
-                        &bind_source_for_table(&reachable.edge_table),
-                        mode,
-                    )
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-            bind_reachable_seed_filters(&mut reachable, binding, schema, mode)?;
-            Ok(reachable)
-        })
-        .collect::<Result<Vec<_>, Error>>()?;
-    query.array_subqueries = query
-        .array_subqueries
-        .into_iter()
-        .map(|subquery| bind_array_subquery_filter_literals(subquery, binding, schema, mode))
-        .collect::<Result<Vec<_>, Error>>()?;
-    query.policy_branches = query
-        .policy_branches
-        .into_iter()
-        .map(|mut branch| {
-            branch.filters = branch
-                .filters
-                .into_iter()
-                .map(|predicate| {
-                    bind_query_predicate(predicate, binding, schema, &root_source, mode)
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-            branch.joins = branch
-                .joins
-                .into_iter()
-                .map(|join| bind_join_filter_literals(join, binding, schema, mode))
-                .collect::<Result<Vec<_>, Error>>()?;
-            branch.reachable = branch
-                .reachable
-                .into_iter()
-                .map(|mut reachable| {
-                    if should_inline_reachable_seed(&reachable.from, mode) {
-                        reachable.from = bind_query_operand(reachable.from, binding, mode)?;
-                    }
-                    reachable.access_filters = reachable
-                        .access_filters
-                        .into_iter()
-                        .map(|predicate| {
-                            bind_query_predicate(
-                                predicate,
-                                binding,
-                                schema,
-                                &bind_source_for_table(&reachable.access_table),
-                                mode,
-                            )
-                        })
-                        .collect::<Result<Vec<_>, _>>()?;
-                    reachable.edge_filters = reachable
-                        .edge_filters
-                        .into_iter()
-                        .map(|predicate| {
-                            bind_query_predicate(
-                                predicate,
-                                binding,
-                                schema,
-                                &bind_source_for_table(&reachable.edge_table),
-                                mode,
-                            )
-                        })
-                        .collect::<Result<Vec<_>, _>>()?;
-                    bind_reachable_seed_filters(&mut reachable, binding, schema, mode)?;
-                    Ok(reachable)
-                })
-                .collect::<Result<Vec<_>, Error>>()?;
-            Ok(branch)
-        })
-        .collect::<Result<Vec<_>, Error>>()?;
-    let rebound = query.validate_with_schema_version(schema, shape.schema_version())?;
-    if rebound.schema_version() != shape.schema_version() {
-        return Err(Error::InvalidStoredValue("bound query schema changed"));
-    }
-    Ok(rebound)
-}
-
-fn bind_array_subquery_filter_literals(
-    mut subquery: ArraySubquery,
-    binding: &Binding,
-    schema: &JazzSchema,
-    mode: ParamBindingMode,
-) -> Result<ArraySubquery, Error> {
-    let source = bind_source_for_table(&subquery.table);
-    subquery.filters = subquery
-        .filters
-        .into_iter()
-        .map(|predicate| bind_query_predicate(predicate, binding, schema, &source, mode))
-        .collect::<Result<Vec<_>, _>>()?;
-    subquery.nested_arrays = subquery
-        .nested_arrays
-        .into_iter()
-        .map(|nested| bind_array_subquery_filter_literals(nested, binding, schema, mode))
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(subquery)
-}
-
-fn inline_snapshot_bind_filter_literals(
-    shape: &ValidatedQuery,
-    binding: &Binding,
-    schema: &JazzSchema,
-) -> Result<ValidatedQuery, Error> {
-    bind_query_params_with_mode(
-        shape,
-        binding,
-        schema,
-        ParamBindingMode::InlineAllReachableSeeds,
-    )
-}
-
-fn retarget_binding_value_sources(shape: &mut NormalizedRowSetShape, binding_source_shape: &str) {
-    for node in shape.nodes.values_mut() {
-        if let RowSetExpr::ValueSource {
-            shape,
-            mode: ValueSourceMode::Binding,
-            ..
-        } = node
-        {
-            *shape = binding_source_shape.to_owned();
-        }
-    }
-}
-
-fn binding_claim_params_for_shape(
-    shape: &NormalizedRowSetShape,
-    param_types: &BTreeMap<String, ColumnType>,
-) -> BTreeMap<String, ProgramClaimParam> {
-    let mut params = BTreeMap::new();
-    for node in shape.nodes.values() {
-        if let RowSetExpr::ValueSource {
-            columns,
-            mode: ValueSourceMode::Binding,
-            ..
-        } = node
-        {
-            for column in columns {
-                let NormalizedValueRef::Claim(path) = &column.value else {
-                    continue;
-                };
-                params.insert(
-                    claim_param_field(path),
-                    ProgramClaimParam {
-                        path: path.clone(),
-                        ty: column.ty.clone(),
-                    },
-                );
-            }
-        }
-        collect_claim_field_params_from_node(node, param_types, &mut params);
-    }
-    params
-}
-
-fn normalized_source_tables(shape: &NormalizedRowSetShape) -> BTreeSet<String> {
-    shape
-        .nodes
-        .values()
-        .filter_map(|node| match node {
-            RowSetExpr::Source { source, .. } => Some(source.table.clone()),
-            _ => None,
-        })
-        .chain(
-            shape
-                .auxiliary_sources
-                .iter()
-                .map(|source| source.table.clone()),
-        )
-        .collect()
-}
-
-fn collect_reachable_seed_claim_params(
-    schema: &JazzSchema,
-    query: &JazzQuery,
-    params: &mut BTreeMap<String, ProgramClaimParam>,
-) -> Result<(), Error> {
-    for reachable in query.reachable.iter().chain(
-        query
-            .policy_branches
-            .iter()
-            .flat_map(|branch| branch.reachable.iter()),
-    ) {
-        let Some(seed) = &reachable.seed else {
-            continue;
-        };
-        let (Some(user_column), Some(user_claim)) = (&seed.user_column, &seed.user_claim) else {
-            continue;
-        };
-        let table = schema
-            .tables
-            .iter()
-            .find(|candidate| candidate.name == seed.table)
-            .ok_or_else(|| Error::TableNotFound(seed.table.clone()))?;
-        let column = table
-            .columns
-            .iter()
-            .find(|candidate| candidate.name == *user_column)
-            .ok_or(Error::InvalidStoredValue(
-                "reachable seed column is missing from schema",
-            ))?;
-        let path = ClaimPath(user_claim.split('.').map(str::to_owned).collect());
-        params.insert(
-            claim_param_field(&path),
-            ProgramClaimParam {
-                path,
-                ty: column.column_type.clone(),
-            },
-        );
-    }
-    Ok(())
-}
-
-fn collect_claim_field_params_from_node(
-    node: &RowSetExpr,
-    param_types: &BTreeMap<String, ColumnType>,
-    params: &mut BTreeMap<String, ProgramClaimParam>,
-) {
-    match node {
-        RowSetExpr::Filter { predicate, .. } | RowSetExpr::Join { on: predicate, .. } => {
-            collect_claim_field_params_from_predicate(predicate, param_types, params);
-        }
-        RowSetExpr::RecursiveRelation {
-            frontier_key,
-            dedupe_keys,
-            ..
-        } => {
-            collect_claim_field_param_authoritative(frontier_key, ColumnType::Uuid, params);
-            for key in dedupe_keys {
-                collect_claim_field_param_authoritative(key, ColumnType::Uuid, params);
-            }
-        }
-        RowSetExpr::Project { columns, .. } => {
-            for column in columns {
-                collect_claim_field_param_authoritative(
-                    &column.value,
-                    column.output.ty.clone(),
-                    params,
-                );
-            }
-        }
-        RowSetExpr::Distinct { keys, .. } => {
-            for key in keys {
-                collect_claim_field_param_authoritative(key, ColumnType::Uuid, params);
-            }
-        }
-        RowSetExpr::CorrelatedPathProjection { correlation, .. } => {
-            collect_claim_field_params_from_predicate(correlation, param_types, params);
-        }
-        RowSetExpr::OrderBy { keys, .. } => {
-            for key in keys {
-                collect_claim_field_param_authoritative(&key.value, ColumnType::Uuid, params);
-            }
-        }
-        RowSetExpr::Slice {
-            partition_by,
-            tie_breaker,
-            ..
-        } => {
-            for value in partition_by.iter().chain(tie_breaker) {
-                collect_claim_field_param_authoritative(value, ColumnType::Uuid, params);
-            }
-        }
-        RowSetExpr::Aggregate {
-            group_by, outputs, ..
-        } => {
-            for value in group_by {
-                collect_claim_field_param_authoritative(value, ColumnType::Uuid, params);
-            }
-            for output in outputs {
-                if let Some(input) = &output.input {
-                    collect_claim_field_param_authoritative(
-                        input,
-                        output.output.ty.clone(),
-                        params,
-                    );
-                }
-            }
-        }
-        RowSetExpr::ValueSource { .. }
-        | RowSetExpr::FrontierSource { .. }
-        | RowSetExpr::Source { .. }
-        | RowSetExpr::Union { .. } => {}
-    }
-}
-
-fn collect_claim_field_params_from_predicate(
-    predicate: &NormalizedPredicateExpr,
-    param_types: &BTreeMap<String, ColumnType>,
-    params: &mut BTreeMap<String, ProgramClaimParam>,
-) {
-    match predicate {
-        NormalizedPredicateExpr::True | NormalizedPredicateExpr::False => {}
-        NormalizedPredicateExpr::Compare { left, right, .. } => {
-            collect_claim_field_param(left, param_types, params);
-            collect_claim_field_param(right, param_types, params);
-        }
-        NormalizedPredicateExpr::In { value, options } => {
-            collect_claim_field_param(value, param_types, params);
-            for option in options {
-                collect_claim_field_param(option, param_types, params);
-            }
-        }
-        NormalizedPredicateExpr::ArrayContains { value, needle }
-        | NormalizedPredicateExpr::TextContains { value, needle } => {
-            collect_claim_field_param(value, param_types, params);
-            collect_claim_field_param(needle, param_types, params);
-        }
-        NormalizedPredicateExpr::IsNull(value) | NormalizedPredicateExpr::IsNotNull(value) => {
-            collect_claim_field_param(value, param_types, params);
-        }
-        NormalizedPredicateExpr::And(children) | NormalizedPredicateExpr::Or(children) => {
-            for child in children {
-                collect_claim_field_params_from_predicate(child, param_types, params);
-            }
-        }
-        NormalizedPredicateExpr::Not(child) => {
-            collect_claim_field_params_from_predicate(child, param_types, params);
-        }
-        // Payload fields belong to the enum value rather than the containing
-        // record. They can still contain claim parameters, so walk the nested
-        // predicate while only collecting the enclosing record value here.
-        NormalizedPredicateExpr::EnumMatch { value, payload, .. } => {
-            collect_claim_field_param(value, param_types, params);
-            collect_claim_field_params_from_predicate(payload, param_types, params);
-        }
-    }
-}
-
-fn collect_claim_field_param(
-    value: &NormalizedValueRef,
-    param_types: &BTreeMap<String, ColumnType>,
-    params: &mut BTreeMap<String, ProgramClaimParam>,
-) {
-    let NormalizedValueRef::Param(param) = value else {
-        return;
-    };
-    let Some(path) = claim_path_from_param_field(param) else {
-        return;
-    };
-    let Some(ty) = param_types.get(param).cloned() else {
-        return;
-    };
-    params
-        .entry(param.clone())
-        .or_insert(ProgramClaimParam { path, ty });
-}
-
-fn collect_claim_field_param_authoritative(
-    value: &NormalizedValueRef,
-    ty: ColumnType,
-    params: &mut BTreeMap<String, ProgramClaimParam>,
-) {
-    let NormalizedValueRef::Param(param) = value else {
-        return;
-    };
-    let Some(path) = claim_path_from_param_field(param) else {
-        return;
-    };
-    params.insert(param.clone(), ProgramClaimParam { path, ty });
-}
-
-fn bind_query_predicate(
-    predicate: Predicate,
-    binding: &Binding,
-    schema: &JazzSchema,
-    source: &SourceId,
-    mode: ParamBindingMode,
-) -> Result<Predicate, Error> {
-    Ok(match predicate {
-        Predicate::All(predicates) => Predicate::All(
-            predicates
-                .into_iter()
-                .map(|predicate| bind_query_predicate(predicate, binding, schema, source, mode))
-                .collect::<Result<Vec<_>, _>>()?,
-        ),
-        Predicate::Any(predicates) => Predicate::Any(
-            predicates
-                .into_iter()
-                .map(|predicate| bind_query_predicate(predicate, binding, schema, source, mode))
-                .collect::<Result<Vec<_>, _>>()?,
-        ),
-        Predicate::Not(predicate) => Predicate::Not(Box::new(bind_query_predicate(
-            *predicate, binding, schema, source, mode,
-        )?)),
-        Predicate::Eq(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Eq)?
-        }
-        Predicate::Ne(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Ne)?
-        }
-        Predicate::In(left, values) => {
-            let left = bind_query_operand(left, binding, mode)?;
-            let target_type = operand_column_type(schema, source, &left)?;
-            Predicate::In(
-                left,
-                values
-                    .into_iter()
-                    .map(|operand| {
-                        bind_query_operand_with_target_type(
-                            operand,
-                            binding,
-                            target_type.as_ref(),
-                            mode,
-                        )
-                    })
-                    .collect::<Result<Vec<_>, _>>()?,
-            )
-        }
-        Predicate::Gt(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Gt)?
-        }
-        Predicate::Gte(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Gte)?
-        }
-        Predicate::Lt(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Lt)?
-        }
-        Predicate::Lte(left, right) => {
-            bind_binary_predicate(left, right, binding, schema, source, mode, Predicate::Lte)?
-        }
-        Predicate::Contains(left, right) => {
-            let left = bind_query_operand(left, binding, mode)?;
-            let needle_type = contains_needle_type(schema, source, &left)?;
-            let right =
-                bind_query_operand_with_target_type(right, binding, needle_type.as_ref(), mode)?;
-            match left {
-                Operand::Literal(Value::Array(values)) => {
-                    let target_type = operand_column_type(schema, source, &right)?;
-                    Predicate::In(
-                        right,
-                        values
-                            .into_iter()
-                            .map(|value| {
-                                Operand::Literal(
-                                    target_type
-                                        .as_ref()
-                                        .map(|target_type| {
-                                            coerce_literal_for_column_type(
-                                                value.clone(),
-                                                target_type,
-                                            )
-                                        })
-                                        .unwrap_or(value),
-                                )
-                            })
-                            .collect(),
-                    )
-                }
-                left => Predicate::Contains(left, right),
-            }
-        }
-        Predicate::IsNull(operand) => {
-            Predicate::IsNull(bind_query_operand(operand, binding, mode)?)
-        }
-        Predicate::EnumMatch {
-            column,
-            case,
-            payload,
-        } => Predicate::EnumMatch {
-            column,
-            case,
-            payload,
-        },
-    })
-}
-
-fn bind_reachable_seed_filters(
-    reachable: &mut crate::query::ReachableVia,
-    binding: &Binding,
-    schema: &JazzSchema,
-    mode: ParamBindingMode,
-) -> Result<(), Error> {
-    if let Some(seed) = &mut reachable.seed {
-        let source = bind_source_for_table(&seed.table);
-        seed.filters = std::mem::take(&mut seed.filters)
-            .into_iter()
-            .map(|predicate| bind_query_predicate(predicate, binding, schema, &source, mode))
-            .collect::<Result<Vec<_>, _>>()?;
-    }
-    Ok(())
-}
-
-fn bind_join_filter_literals(
-    mut join: JoinVia,
-    binding: &Binding,
-    schema: &JazzSchema,
-    mode: ParamBindingMode,
-) -> Result<JoinVia, Error> {
-    let source = bind_source_for_table(&join.table);
-    join.filters = join
-        .filters
-        .into_iter()
-        .map(|predicate| bind_query_predicate(predicate, binding, schema, &source, mode))
-        .collect::<Result<Vec<_>, _>>()?;
-    join.nested_joins = join
-        .nested_joins
-        .into_iter()
-        .map(|join| bind_join_filter_literals(join, binding, schema, mode))
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(join)
-}
-
-fn bind_binary_predicate(
-    left: Operand,
-    right: Operand,
-    binding: &Binding,
-    schema: &JazzSchema,
-    source: &SourceId,
-    mode: ParamBindingMode,
-    build: impl FnOnce(Operand, Operand) -> Predicate,
-) -> Result<Predicate, Error> {
-    let left_type = operand_column_type(schema, source, &left)?;
-    let right_type = operand_column_type(schema, source, &right)?;
-    Ok(build(
-        bind_query_operand_with_target_type(left, binding, right_type.as_ref(), mode)?,
-        bind_query_operand_with_target_type(right, binding, left_type.as_ref(), mode)?,
-    ))
-}
-
-fn bind_source_for_table(table: &str) -> SourceId {
-    SourceId {
-        table: table.to_owned(),
-        path: SourcePath {
-            components: Vec::new(),
-        },
-    }
-}
-
-fn should_inline_reachable_seed(operand: &Operand, mode: ParamBindingMode) -> bool {
-    match (operand, mode) {
-        (Operand::Param(_), ParamBindingMode::InlineAllReachableSeeds) => true,
-        (Operand::Param(_), ParamBindingMode::RetainAllParams) => false,
-        _ => false,
-    }
-}
-
-fn bind_query_operand(
-    operand: Operand,
-    binding: &Binding,
-    mode: ParamBindingMode,
-) -> Result<Operand, Error> {
-    bind_query_operand_with_target_type(operand, binding, None, mode)
-}
-
-fn bind_query_operand_with_target_type(
-    operand: Operand,
-    binding: &Binding,
-    target_type: Option<&ColumnType>,
-    mode: ParamBindingMode,
-) -> Result<Operand, Error> {
-    Ok(match operand {
-        Operand::Param(name) if matches!(mode, ParamBindingMode::RetainAllParams) => {
-            Operand::Param(name)
-        }
-        Operand::Param(name) => {
-            let value = binding
-                .values()
-                .get(&name)
-                .cloned()
-                .ok_or_else(|| QueryError::MissingParam(name.clone()))?;
-            Operand::Literal(
-                target_type
-                    .map(|target_type| coerce_literal_for_column_type(value.clone(), target_type))
-                    .unwrap_or(value),
-            )
-        }
-        Operand::Literal(value) => Operand::Literal(
-            target_type
-                .map(|target_type| coerce_literal_for_column_type(value.clone(), target_type))
-                .unwrap_or(value),
-        ),
-        Operand::Column(_) | Operand::Claim(_) => operand,
-    })
-}
-
+use bindings::*;
 fn local_maintained_view_content_witness<'a>(
     versions: &'a [VersionRow],
     table: &str,
